[Fix] File structure to feature-folder
