+++ client/src/app/rootReducer.js
@@ -1,0 +1,5 @@
+import { combineReducers } from 'redux';
+
+export default combineReducers({
+    // state here
+});
