+++ client/src/app/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+import rootReducer from './rootReducer';
+const store = configureStore({
+    reducer: rootReducer
+});
+
+export default store;
