+++ client/src/components/CardsContainer/img/course.jpeg
