+++ client/src/components/FadeText/FadeText.css
