+++ client/src/components/FadeText/index.js
