+++ client/src/components/Landing/Goals.js
@@ -5,7 +5,7 @@ const Goals = () => (
         <h2>Achieve your goals with Study Tube</h2>
         <div className='cards'>
             <div className='card'>
-                <img src={require('./img/skill.png')} alt='Skill icon' />
+                <img src={require('../../img/skill.png')} alt='Skill icon' />
                 <h3>Learn the latest skills</h3>
                 <p>
                     Lorem ipsum dolor sit amet, consectetur adipiscing elit. In
@@ -13,7 +13,7 @@ const Goals = () => (
                 </p>
             </div>
             <div className='card'>
-                <img src={require('./img/career.png')} alt='Career icon' />
+                <img src={require('../../img/career.png')} alt='Career icon' />
                 <h3>Get ready for a career</h3>
                 <p>
                     Lorem ipsum dolor sit amet, consectetur adipiscing elit. In
@@ -22,7 +22,7 @@ const Goals = () => (
             </div>
             <div className='card'>
                 <img
-                    src={require('./img/certificate.png')}
+                    src={require('../../img/certificate.png')}
                     alt='Certificate icon'
                 />
                 <h3>Earn a certificate</h3>
@@ -33,7 +33,7 @@ const Goals = () => (
             </div>
             <div className='card'>
                 <img
-                    src={require('./img/organization.png')}
+                    src={require('../../img/organization.png')}
                     alt='Organization card'
                 />
                 <h3>Upskill your organization</h3>
