+++ client/src/features/Course/Content.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import Collapse from '../layout/Collapse';
+import Collapse from '../../components/Collapse';
 
 const Content = props => {
     const { editing } = props;
