+++ client/src/features/Course/Feedback.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Rating from '../layout/Rating';
+import Rating from '../../components/Rating';
 
 const Feedback = () => (
     <section>
